Extract the NewsletterSignup in its own file
