+++ packages/js/src/workouts/components/ConfigurationWorkout.js
@@ -15,105 +15,13 @@ import { OrganizationSection } from "./OrganizationSection";
 import { PersonSection } from "./PersonSection";
 import { SocialInput } from "./SocialInput";
 import PropTypes from "prop-types";
+import { NewsletterSignup } from "./NewsletterSignup";
 
 window.wpseoScriptData = window.wpseoScriptData || {};
 window.wpseoScriptData.searchAppearance = {
 	...window.wpseoScriptData.searchAppearance,
 	userEditUrl: "/wp-admin/user-edit.php?user_id={user_id}",
 };
-
-/**
- * A function to request a sign up to the newsletter.
- *
- * @param {string} email The email to signup to the newsletter.
- *
- * @returns {Object} The request's response.
- */
-async function postSignUp( email ) {
-	const response = await fetch( "https://my.yoast.com/api/Mailing-list/subscribe", {
-		method: "POST",
-		mode: "cors",
-		cache: "no-cache",
-		credentials: "same-origin",
-		headers: {
-			"Content-Type": "application/json",
-		},
-		redirect: "follow",
-		referrerPolicy: "no-referrer",
-		body: JSON.stringify(
-			{
-				customerDetails: {
-					firstName: "",
-					email,
-				},
-				list: "Yoast newsletter",
-			}
-		),
-	} );
-	return response.json();
-}
-
-/**
- * The newsletter signup section.
- *
- * @returns {WPElement} A newslettersignup element.
- */
-function NewsletterSignup() {
-	const [ newsletterEmail, setNewsletterEmail ] = useState( "" );
-	const [ signUpState, setSignUpState ] = useState( "ready" );
-
-	const onSignUpClick = useCallback(
-		async function() {
-			setSignUpState( "loading" );
-			await postSignUp( newsletterEmail );
-			setSignUpState( "ready" );
-		},
-		[ newsletterEmail ]
-	);
-
-	return (
-		<>
-			<ul className="yoast-list--usp">
-				<li>{ __( "Receive best-practice tips and learn how to rank on search engines", "wordpress-seo" ) }</li>
-				<li>{ __( "Stay up-to-date with the latest SEO news", "wordpress-seo" ) }</li>
-				<li>{ __( "Get guidance on how to use Yoast SEO to the fullest", "wordpress-seo" ) }</li>
-			</ul>
-			<div className="yoast-newsletter-signup">
-				<TextInput
-					label={ __( "Email address", "wordpress-seo" ) }
-					id="newsletter-email"
-					name="newsletter email"
-					value={ newsletterEmail }
-					onChange={ setNewsletterEmail }
-					type="email"
-				/>
-				<Button
-					variant="primary"
-					onClick={ onSignUpClick }
-					disabled={ signUpState === "loading" }
-				>
-					{ __( "Sign up!", "wordpress-seo" ) }
-				</Button>
-			</div>
-			<p className="yoast-privacy-policy">
-				{
-					addLinkToString(
-						sprintf(
-							// translators: %1$s and %2$s are replaced by opening and closing anchor tags.
-							__(
-								"Yoast respects your privacy. Read %1$sour privacy policy%2$s on how we handle your personal information.",
-								"wordpress-seo"
-							),
-							"<a>",
-							"</a>"
-						),
-						"https://yoa.st/gdpr-config-workout"
-					)
-				}
-			</p>
-		</>
-	);
-}
 
 /* eslint-disable complexity */
 /**
